functions-python/requirements.txt
# functions-python/requirements.txt

# Required Firebase libraries
firebase-admin

# Web framework (dependency for HTTP functions)
Flask

# For making external HTTP requests (CalAPI, Magisterium)
requests

# For Google Generative AI (Gemini) API calls
google-generativeai

# Google API Core libraries
google-api-core

# For handling Cross-Origin Resource Sharing (CORS)
flask-cors

# Production WSGI server
gunicorn